Add loader for v3
