+++ src/main/java/org/entur/gbfs/Version.java
@@ -1,0 +1,6 @@
+package org.entur.gbfs;
+
+public enum Version {
+    V2,
+    V3
+}
